feat(transactions): update row changes when reverting to original cell value
